refact/Delete legacy code and solve some flake8 warnings (#48)
* Remove legacy code and resolve some flake8 warnings

* fix small error
