+++ hummingbot/connector/exchange/idex/__init__.py
@@ -1,0 +1,11 @@
+"""Idex exchange connector for Hummingbot trading bot
+
+Provides a connector for executing Hummingbot's strategies on Idex exchange.
+Exchange website: https://idex.io/
+API documentation: https://docs.idex.io/
+"""
+
+__author__ = "Brian Morenus, Alfredo Valles Valdes, and Elliott Williams"
+__maintainer__ = "Alfredo Valles Valdes"
+__status__ = "Production"
+__version__ = "1.0.0"
